requirements.txt
fastapi[standard]
pymongo
python-dotenv
requests